most funny keys work now, but something is still wrong with numlock

+++ keycodes.h
@@ -9,7 +9,7 @@ extern "C" {
 #endif
 
 // map a PS/2 key code to USB. returns 0 if there is no mapping
-uint8_t ps2_to_usb_keycode(uint8_t);
+uint16_t ps2_to_usb_keycode(uint8_t);
 
 #ifdef __cplusplus 
 } // end of extern "C"
